+++ .babelrc
@@ -1,0 +1,4 @@
+{
+  "presets" : ["env", "react"]
+
+}
