+++ tests/test.js
@@ -2,3 +2,7 @@
 test('adds 1 + 2 to equal 3', () => {
   expect(1 + 2).toBe(3);
 });
+
+
+//name t=file ...app.test.js  import component and react
+
